Query System Overhaul
Any part of the construction of the data returned by a query should not
be done outside of BackendHandler. Also, to make the method name more
consistent with its behavior, the getQueryResponse() method in
BackendHandler has been renamed to getDataFromQuery() and returns
a generic object. Now it will return whatever data was being requested
instead of InformationListQueryResponse or InformationQueryResponse.
These classes, along with QueryResponse have been removed.