Added Offline Mode
When the client is unable to contact the backend, it is put into a
limited state. This way you are still able to manage the programs
that are installed but some features are unavailable.

DataHandler::getCategories() doesn't need a boolean anymore and will
return the latest categories, and if unavailable, will return the
installed categories. Installed programs are now the first to be added
to the list returned from DataHandler::getPrograms() and any new
programs from the backend are added to it, if available.