Client Update Behavior Change
The behavior now is to only check for updates when the setting is
enabled, which it is by default. The previous behavior was to
initiate an update request, and only show notifications of an
update if the setting is enabled. This was intended to be used
as part of the statistics system. However, due to a recent
decision not to have a statistics system, at least for the
foreseeable future, this behavior was changed to match a
previous implementation.

The setting "Show Client Update Notification On Startup" has
been changed to "Check Client Update On Startup" as a result