Date And Time Format Changes
Due to a change in the backend, all date and time values are now
presented in the ISO 8601 instant format. As a result, the
Timestamp class has been replaced by the Instant class.

The updater has also been updated to account for this change as
well as introduce some other changes:

- Version number added to updater releases, now supplied in the user agent
- Handles the new ISO 8601 instant format
- Fixed date and time format when creating error log file